test: switch pnpm version
